+++ CHANGES.rst
@@ -18,6 +18,8 @@
 
 - Drop support for pg8000 older than 1.11.0.
 
+- Drop support for MySQL Connector/Python older than 8.0.16.
+
 - Drop support for PyPy older than 5.3.1.
 
 - Drop support for the "MySQL Connector/Python" driver name since it